refactor(inputs): Move input code to new `core::inputs` mod
