+++ src/network/inputs.rs
@@ -1,6 +1,9 @@
 use crate::{
-    core::physics::PhysicsSet,
-    entities::player::{Player, PlayerAction},
+    core::{
+        inputs::{PlayerAction, PlayerActionState},
+        physics::PhysicsSet,
+    },
+    entities::player::Player,
     network::SessionConfig,
 };
 use bevy::{platform::collections::HashMap, prelude::*};
@@ -30,16 +33,7 @@ const INPUT_RELOAD: u32 = 1 << 11;
 pub struct NetworkInputsPlugin;
 impl Plugin for NetworkInputsPlugin {
     fn build(&self, app: &mut App) {
-        app.add_systems(
-            ReadInputs,
-            (
-                update_gamepad_shoot_input, // Running it before `ulpd` ensures it only affects shoot input if a gamepad acts on `ShootDirection`
-                update_local_pointer_direction,
-                read_local_inputs,
-            )
-                .chain(),
-        )
-        .add_systems(
+        app.add_systems(ReadInputs, read_local_inputs).add_systems(
             GgrsSchedule,
             update_remote_inputs.before(PhysicsSet::Player),
         );
@@ -48,7 +42,7 @@ impl Plugin for NetworkInputsPlugin {
 
 fn read_local_inputs(
     mut commands: Commands,
-    query: Query<(&Player, &ActionState<PlayerAction>)>,
+    query: Query<(&Player, &PlayerActionState)>,
     local_players: Res<LocalPlayers>,
 ) {
     let mut local_inputs = HashMap::new();
@@ -69,12 +63,12 @@ fn read_local_inputs(
 }
 
 fn update_remote_inputs(
-    mut query: Query<(&Player, &mut ActionState<PlayerAction>)>, // Don't filter out LocalPlayer as we don't want his outputs to fire early
+    mut query: Query<(&Player, &mut PlayerActionState)>, // Don't filter out LocalPlayer as we don't want his outputs to fire early
     inputs: Res<PlayerInputs<SessionConfig>>,
 ) {
     for (player, mut action_state) in query.iter_mut() {
         let (input, _) = inputs[player.handle];
-        *action_state = ActionState::<PlayerAction>::from_ggrs_session_input(input);
+        *action_state = PlayerActionState::from_ggrs_session_input(input);
     }
 }
 
@@ -84,7 +78,7 @@ pub trait GgrsSessionInput {
     fn from_ggrs_session_input(input: NetworkInputs) -> Self;
 }
 
-impl GgrsSessionInput for ActionState<PlayerAction> {
+impl GgrsSessionInput for PlayerActionState {
     fn as_ggrs_session_input(&self) -> NetworkInputs {
         let mut keys = 0;
 
@@ -162,49 +156,5 @@ impl GgrsSessionInput for ActionState<PlayerAction> {
         action_state.set_axis_pair(&PlayerAction::PointerDirection, input.pointer_direction);
 
         action_state
-    }
-}
-
-fn update_local_pointer_direction(
-    mut player_query: Query<(&Player, &GlobalTransform, &mut ActionState<PlayerAction>)>,
-    windows: Query<&Window>,
-    query_view: Query<(&Camera, &GlobalTransform), With<Camera2d>>,
-    local_players: Res<LocalPlayers>,
-) -> Result {
-    let window = windows.single()?;
-    let (camera, view) = query_view.single()?;
-    if let Some(world_position) = window
-        .cursor_position()
-        .map(|cursor| camera.viewport_to_world_2d(view, cursor).unwrap())
-    {
-        for (_, player_world_pos, mut action_state) in
-            player_query.iter_mut().filter(|(player, _, action)| {
-                local_players.0.contains(&player.handle)
-                    && action
-                        .dual_axis_data(&PlayerAction::PointerDirection)
-                        .is_none()
-            })
-        {
-            let pointer_direction =
-                (world_position - player_world_pos.translation().xy()).normalize();
-
-            action_state.set_axis_pair(&PlayerAction::PointerDirection, pointer_direction);
-        }
-    } else {
-        // Not an error pointer could be out of window
-    }
-
-    Ok(())
-}
-
-fn update_gamepad_shoot_input(mut query: Query<&mut ActionState<PlayerAction>>) {
-    for mut action_state in query.iter_mut() {
-        if action_state
-            .axis_pair(&PlayerAction::PointerDirection)
-            .length()
-            > 0.6
-        {
-            action_state.set_button_value(&PlayerAction::Shoot, 1.0);
-        }
     }
 }
